refactor(credentials): extract keyring_stub tests to sibling file
